pyproject.toml
[tool.poetry]
name = "kurs-7-dz-24-1"
version = "0.1.0"
description = ""
authors = ["leonid Petrov <leonidpetrov637@gmail.com>"]
readme = "README.md"
packages = [{include = "kurs_7_dz_24"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
black = "^24.4.2"
pillow = "^10.4.0"
isort = "^5.13.2"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
ipython = "^8.26.0"
coverage = "^7.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
